Remove legend state from quantity input
